+++ Sources/DTCollectionViewManager/DTCollectionViewManagerAnomalyHandler.swift
@@ -34,7 +34,10 @@ public enum DTCollectionViewManagerAnomaly: Equatable, CustomStringConvertible, 
     case nilSupplementaryModel(kind: String, indexPath: IndexPath)
     case noCellMappingFound(modelDescription: String, indexPath: IndexPath)
     case noSupplementaryMappingFound(modelDescription: String, kind: String, indexPath: IndexPath)
+    
+    @available(*, deprecated, message: "If you are using xibs or code cells, you should not set reuseIdentifier, because UICollectionView.CellRegistration gives you a random one. If you use storyboard, set reuseIdentifier equal to name of the cell subclass.")
     case differentCellReuseIdentifier(mappingReuseIdentifier: String, cellReuseIdentifier: String)
+    
     case differentSupplementaryReuseIdentifier(mappingReuseIdentifier: String, supplementaryReuseIdentifier: String)
     case differentCellClass(xibName: String, cellClass: String, expectedCellClass: String)
     case differentSupplementaryClass(xibName: String, viewClass: String, expectedViewClass: String)